Adds extra information to readme and lowers update_frequency in dev
This patch updates the readme to add additional configuration information that
may help the getting started experience for developers and a small configuration
change to the development config.

Upon implementing elixometer I struggled to figure out why my reporter wasn't being
called, neither a custom reporter nor :exometer_report_tty was working.

After routing around in the source I found that you must register your reporter with
both exometer AND elixometer in order for it to work correctly.

This patch hopes to share that knowledge with the community.