Addressing bottleneck
In one of our high-throughput systems, timing messages were a bottleneck
because the update calls were synchronous. This PR moves the calls to an
updater process that is capped at 1000 messages in its queue using the
pobox framework. This should alleviate the bottleneck and prevent the
process mailboxes from getting too big.