Test our oldest and newest Elixir version pairs (#65)
This revises the build matrix so that we test the oldest supported
Elixir version (now 1.3.0) as well as the newest (currently 1.4.0).

Note that increase in minimum version to Elixir 1.3.0. Due to previous
changes related to our lager version override, we no longer seem able
to support Elixir 1.0.0.

This also removes some unnecessary setup steps; all of the removed
things already happen by default.